list_model_classes

Returns the possible model classes that `gglm` works with.

Usage

list_model_classes(...)

Arguments

...

Currently ignored. For extendability.

Value

A character vector containing the possible model classes that `gglm` works with.

Note

Note that these are not always the exact name of the class that that can be used. This is due to how some methods are written in the packages `gglm` imports. For example, the model class "merMod" refers to a variety of model outputs from `lme4`, even when the outputted class is not called "merMod".

Examples

list_model_classes()
#>  [1] "mlogit"        "rq"            "logitmfx"      "Mclust"       
#>  [5] "betamfx"       "poLCA"         "clmm"          "nls"          
#>  [9] "factanal"      "poissonmfx"    "htest"         "default"      
#> [13] "lmrob"         "brmsfit"       "prcomp"        "smooth.spline"
#> [17] "polr"          "glmRob"        "stl"           "mfx"          
#> [21] "felm"          "glmmTMB"       "glmmadmb"      "survreg"      
#> [25] "drc"           "fixest"        "clm"           "lme"          
#> [29] "pam"           "lm"            "glmrob"        "decomposed.ts"
#> [33] "rqs"           "gam"           "loess"         "probitmfx"    
#> [37] "negbinmfx"     "ranef.mer"     "kmeans"        "merMod"       
#> [41] "nlrq"          "rlm"           "margins"       "rma"          
#> [45] "ivreg"         "mjoint"        "sarlm"         "data.frame"   
#> [49] "speedglm"      "plm"           "coxph"         "glm"          
#> [53] "lmRob"         "betareg"       "NULL"          "gamm4"        
#> [57] "speedlm"       "Sarlm"         "gls"